:recycle: Move DefaultService model in code
